+++ CHANGELOG.md
@@ -5,8 +5,12 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Added
+- Added headers to publisher
 ### Change
 - Default timeout from 5 to 0
+### Removed
+- Remove `$properties` from `PublisherContract::publish()`, `PublisherContract::rpc()`
 
 ## [v1.0.0-alpha-1]
 ### Fixed